DRY refactoring
