+++ lib/cachify.js
@@ -21,7 +21,7 @@ function parseOptions(options) {
 
 function cachify(func, options) {
 	if (typeof func !== 'function') {
-		return func;
+		throw new Error("Must be a function");
 	}
 
 	options = parseOptions(options);
@@ -42,27 +42,10 @@ function cachifyAll(target, options) {
 	if (typeof options.context !== 'object') {
 		options.context = target;
 	}
-
-	_.each(target, function(fn, fnName) {
-		
-		var passesDefaultFilter = options.filter === util.defaultFilter
-		    ? true : util.defaultFilter(fnName);
-
-		if (typeof options.filter === 'function') {
-			var passes = options.filter(fnName, fn, target, passesDefaultFilter);
-			if (!passes) {
-				return;
-			}
-		}
-
+	
+	util.forEachFilteredFunction(target, options, function(fn, fnName) {
 		var newFnName = fnName + options.suffix;
-		if (newFnName !== fnName && target[newFnName]) {
-			throw new Error("Property already exists '" + newFnName + "'");
-		}
-
-		target[newFnName] = cachify(fn, _.extend({}, options, {
-			keyPrefix: newFnName
-		}));
+		target[newFnName] = cachify(fn, options);
 	});
 
 	return target;
