+++ lib/util.js
@@ -1,5 +1,7 @@
 /* jshint node: true */
 
+var _ = require('@dispatcher/underscore-ext');
+
 var rident = /^[a-z$_][a-z$_0-9]*$/i;
 
 function isIdentifier(str) {
@@ -12,5 +14,33 @@ function defaultFilter(name) {
         name !== "constructor";
 };
 
+function forEachFilteredFunction(target, options, callback) {
+
+	_.each(target, function(fn, fnName) {
+		
+		if (typeof fn !== 'function') {
+			return;
+		}
+
+		var passesDefaultFilter = options.filter === defaultFilter
+		    ? true : defaultFilter(fnName);
+
+		if (typeof options.filter === 'function') {
+			var passes = options.filter(fnName, fn, target, passesDefaultFilter);
+			if (!passes) {
+				return;
+			}
+		}
+
+		var newFnName = fnName + options.suffix;
+		if (newFnName !== fnName && target[newFnName]) {
+			throw new Error("Property already exists '" + newFnName + "'");
+		}
+
+		callback(fn, fnName);
+	});
+}
+
 module.exports.isIdentifier = isIdentifier;
 module.exports.defaultFilter = defaultFilter;
+module.exports.forEachFilteredFunction = forEachFilteredFunction;
